requirements.txt
fastapi
uvicorn
discord.py
jinja2
supabase
python-dotenv
httpx
python-multipart
